sync

+++ front_back_interaction/chats/backend/src/person/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 class Student(models.Model):
     user = models.OneToOneField(User,on_delete=models.CASCADE)
     nickname = models.CharField(max_length=20)
-    progress = models.FloatField(max_digits=5, decimal_places=2)
+    progress = models.DecimalField(max_digits=5, decimal_places=2)
